mobihive/requirements.txt
django
djangorestframework
Pillow
django-cors-headers
djangorestframework-simplejwt